services/worker/requirements.txt
celery==5.4.0
redis==5.1.1
pymongo==4.10.1
boto3==1.35.36
python-dotenv==1.0.1
pydantic-settings==2.5.2
pytesseract==0.3.13
pdf2image==1.17.0
Pillow==10.4.0
